# server/src/main/openapi/fintube.yaml
openapi: 3.0.3
info: { title: FinTube API, version: 1.0.0 }
servers: [ { url: /api } ]
paths:
...
  /me:
    get: { tags: [ Account ], operationId: getCurrentUser, responses: { '200': { description: Current user, content: { application/json: { schema: { $ref: '#/components/schemas/Profile' } } } }, '401': { description: Unauthenticated } } }
  /me/youtube-api-key:
    get: { tags: [ Account ], operationId: getYouTubeApiKeyStatus, responses: { '200': { description: YouTube API key status, content: { application/json: { schema: { $ref: '#/components/schemas/YouTubeApiKeyStatus' } } } }, '401': { description: Unauthenticated } } }
    put: { tags: [ Account ], operationId: updateYouTubeApiKey, requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateYouTubeApiKeyRequest' } } } }, responses: { '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/YouTubeApiKeyStatus' } } } }, '400': { description: Invalid key }, '401': { description: Unauthenticated } } }
...
  /me/change-password:
    post: { tags: [ Account ], operationId: changePassword, requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ChangePasswordRequest' } } } }, responses: { '200': { description: Changed } } }
...
components:
  schemas:
    Role: { type: string, enum: [ USER, ADMIN ] }
...
    Profile: { type: object, required: [ id, username, role, filesystemSlug ], properties: { id: { type: integer, format: int64 }, username: { type: string }, role: { $ref: '#/components/schemas/Role' }, filesystemSlug: { type: string } } }
    YouTubeApiKeyStatus: { type: object, required: [ configured ], properties: { configured: { type: boolean } } }
    UpdateYouTubeApiKeyRequest: { type: object, required: [ api_key ], properties: { api_key: { type: string, maxLength: 4096 } } }
...
    ChangePasswordRequest: { type: object, required: [ oldPassword, newPassword ], properties: { oldPassword: { type: string }, newPassword: { type: string, minLength: 6 } } }
...